move python compilation path into gtc2.python
