+++ src/gtc2/python/pnir_to_ast.py
@@ -11,8 +11,8 @@ from gt4py.backend.gtc_backend.stencil_module_builder import (
     StencilModuleBuilder,
     parse_node,
 )
-
-from . import common, gtir, pnir
+from gtc2 import common, gtir
+from gtc2.python import pnir
 
 
 GET_DOMAIN_TPL = """
