+++ tests/test_unittest/test_gtc2/test_gtir.py
@@ -18,7 +18,7 @@ from gtc2.gtir import (
     VerticalInterval,
     VerticalLoop,
 )
-from gtc2.python_naive_codegen import PythonNaiveCodegen
+from gtc2.python.python_naive_codegen import PythonNaiveCodegen
 
 
 @pytest.fixture
