+++ tests/test_unittest/test_gtc2/test_gtir_to_pnir.py
@@ -2,8 +2,9 @@ from typing import Iterator
 
 import pytest
 
-from gtc2 import common, gtir, pnir
-from gtc2.gtir_to_pnir import GtirToPnir
+from gtc2 import common, gtir
+from gtc2.python import pnir
+from gtc2.python.gtir_to_pnir import GtirToPnir
 
 
 @pytest.fixture
