+++ tests/test_unittest/test_gtc2/test_pnir_to_ast.py
@@ -8,8 +8,9 @@ import numpy
 import pytest
 
 from gt4py.utils import make_module_from_file
-from gtc2 import common, gtir, pnir
-from gtc2.pnir_to_ast import PnirToAst
+from gtc2 import common, gtir
+from gtc2.python import pnir
+from gtc2.python.pnir_to_ast import PnirToAst
 
 
 def to_snippet(node: ast.AST) -> str:
